Rename, simplify validation functions, module
